# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExampleSms API",
        "description": "This is a sample ExampleSms API server.  You can find out more about Swagger at",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "ExampleSms API v1.0.0 server"
        }
    ],
    "paths": {
...
        "/send-sms": {
            "post": {
                "tags": [
                    "sms"
                ],
                "summary": "Send SMS",
                "description": "Send SMS endpoint.",
                "operationId": "sendSms",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "example": "905*********"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "SMS Sender",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "example": "SMS Sender"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "SMS Message",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "example": "This is message."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "message": "User logged in successfully.",
                                            "data": {
                                                "name": "First Last",
                                                "email": "user@example.com",
                                                "username": "user234",
                                                "address": "1234 Main St",
                                                "phone": "1234567890"
                                            },
                                            "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvYXBpL3YxL2xvZ2luIiwiaWF0IjoxNzA1MTA0Njc3LCJleHAiOjE3MDUxMDgyNzcsIm5iZiI6MTcwNTEwNDY3NywianRpIjoiOGtTR1JLckRLYXFJaHlYbyIsInN1YiI6IjgiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.R_TzWgGlxPHY-tjIN1255GLWB7qdvRY9CZe39p2rixA",
                                            "token_type": "bearer",
                                            "expires_in": 3600
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}